+++ frontend/src/business/hooks/index.ts
@@ -2,4 +2,4 @@ export * from './useBlocker';
 export * from './useOverflowTextBoxCenter';
 export * from './useShareButtons';
 export * from './useSpeakerHighlighter';
-export * from './useTOLD';
+export * from './useUserFeedback';
