+++ frontend/src/business/services/TOLD.ts
@@ -1,0 +1,12 @@
+const TOLD = {
+  init: () => ((window as any).dataLayer = []),
+  displayForm: (type: string) => (window as any).dataLayer.push({ event: type }),
+  isFormVisible: () => {
+    const container = document.querySelector('#told-container');
+    return container?.innerHTML ?? false;
+  },
+};
+
+TOLD.init();
+
+export default TOLD;
